refact: add index_util
