+++ src/schema/index_util.h
@@ -1,0 +1,48 @@
+/*
+ * Copyright 2021 4Paradigm
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef SRC_SCHEMA_INDEX_UTIL_H_
+#define SRC_SCHEMA_INDEX_UTIL_H_
+
+#include "base/status.h"
+#include "proto/common.pb.h"
+#include "proto/name_server.pb.h"
+#include "vm/catalog.h"
+
+namespace openmldb {
+namespace schema {
+
+using PBSchema = ::google::protobuf::RepeatedPtrField<::openmldb::common::ColumnDesc>;
+using PBIndex = ::google::protobuf::RepeatedPtrField<::openmldb::common::ColumnKey>;
+
+class IndexUtil {
+ public:
+    static bool ConvertIndex(const PBIndex& index, ::hybridse::vm::IndexList* output);
+
+    static base::Status CheckIndex(const std::map<std::string, ::openmldb::type::DataType>& column_map,
+            const PBIndex& index);
+
+    static bool CheckTTL(const ::openmldb::common::TTLSt& ttl);
+
+    static bool AddDefaultIndex(TableInfo* table_info);
+
+    static bool FillColumnKey(TableInfo* table_info);
+
+};
+
+}  // namespace schema
+}  // namespace openmldb
+#endif  // SRC_SCHEMA_INDEX_UTIL_H_
