+++ src/sdk/sql_router_test.cc
@@ -27,7 +27,6 @@
 #include "base/file_util.h"
 #include "base/glog_wapper.h"
 #include "case/sql_case.h"
-#include "catalog/schema_adapter.h"
 #include "codec/fe_row_codec.h"
 #include "common/timer.h"
 #include "gflags/gflags.h"
